Special module with boilerplate helpers
